refactor: components라는 컴포넌트에서 다른 도메인 폴더들과 구분하기 위해서 @를 붙임
