+++ frontend/src/components/@layout/Header/Header.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 
 import gongseek from '@/assets/gongseek.png';
-import * as S from '@/components/layout/Header/Header.styles';
+import * as S from '@/components/@layout/Header/Header.styles';
 import SearchBar from '@/components/search/SearchBar/SearchBar';
 import UserProfileIcon from '@/components/user/UserProfileIcon/UserProfileIcon';
 import { URL } from '@/constants/url';
