+++ frontend/src/components/@layout/TabBar/TabBar.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from 'react-router-dom';
 import { useRecoilState, useRecoilValue } from 'recoil';
 
 import gongseek from '@/assets/gongseek.png';
-import * as S from '@/components/layout/TabBar/TabBar.styles';
+import * as S from '@/components/@layout/TabBar/TabBar.styles';
 import UserProfileIcon from '@/components/user/UserProfileIcon/UserProfileIcon';
 import { UserProfile } from '@/components/user/UserProfileIcon/UserProfileIcon.styles';
 import { URL } from '@/constants/url';
