+++ frontend/src/storybook/Layout/Header.stories.tsx
@@ -1,4 +1,4 @@
-import Header from '@/components/layout/Header/Header';
+import Header from '@/components/@layout/Header/Header';
 import { Story, Meta } from '@storybook/react';
 
 export default {
